% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owri_db.R
\name{owri_db}
\alias{owri_db}
\title{Current OWRI Database}
\usage{
owri_db(owri.db = "OWRIDbExport_011022.db")
}
\arguments{
\item{owri.db}{Optional. owri SQLite database file name. Default is the
most recent version. 'OWRIDbExport_011022.db'.}
}
\description{
Returns the path to the most recent SQLite OWRI database export included with this R pacakge.
}
\keyword{owri,}
\keyword{raw_data}
