% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owri_summary20.R
\name{owri_summary20}
\alias{owri_summary20}
\title{Summarize OWRI treatments over 20 years.}
\usage{
owri_summary20(owri.db, complete.years, huc8)
}
\arguments{
\item{owri.db}{The path and file name of the owri SQLite database.}

\item{complete.years}{Vector of numeric years used to fetch projects. The year range must span 20 years. Only projects completed in these years will be summarized.}

\item{huc8}{Vector of numeric HUC8 codes used to fetch projects. Only projects within the HUC8 will be summarized.}
}
\value{
Dataframe with the sum of treatments implemented grouped into four yearly periods by huc8 and activity type.
}
\description{
This function is primarily used for DEQ's Water Quality Status and Trends Report.
}
\keyword{complete}
\keyword{huc8,}
\keyword{owri,}
\keyword{years,}
