% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb_to_sqlite.R
\name{mdb_to_sqlite}
\alias{mdb_to_sqlite}
\title{Move Access Tables into a SQLite Database.}
\usage{
mdb_to_sqlite(mdb, sqlite)
}
\arguments{
\item{mdb}{The full path and file name to the Access database.}

\item{sqlite}{The path and file name to the new SQLite database to be created.}
}
\value{
mdb_to_sqlite() has no returned value.
}
\description{
Move OWRI tables in a Microsoft Access database into a SQlite database.
All tables are moved exactly as named using their existing column datatypes.
If a table already exists in the SQlite database an error will be returned.
This function is to be used to create new tables in a SQlite database rather
than append or overwrite them.
}
\details{
This function must be utilized with the 32 bit version of R. If using RStudio
and converting a *.mdb Microsoft Access database only the 32 bit compatible
version of RStudio Desktop can be used (<= version 1.1.463).
Alternatively if the *.mdb file can be opened with Access and saved as a *.accdb
file, newer versions of RStudio Desktop can be used instead.

Microsoft Access 1997 or newer supported with Microsoft Access Driver (*.mdb, *.accdb) installed.
}
\keyword{access,}
\keyword{database}
\keyword{mdb,}
\keyword{sqlite,}
