% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb_to_sqlite.R
\name{mdb_to_sqlite}
\alias{mdb_to_sqlite}
\title{Move Access mdb Tables into a SQLite Database.}
\usage{
mdb_to_sqlite(mdb, sqlite)
}
\arguments{
\item{mdb}{The path and file name to the Access database.}

\item{sqlite}{The path and file name to the new SQLite database to be created.
columns and datatypes are the same as a dataframe returned from AWQMSdata::AWQMS_data(). Defualt is NULL.}

\item{...}{Other arguments passed on to methods.}
}
\value{
mdb_to_sqlite() has no returned value.
}
\description{
Move tables in a Microsoft Access database into a SQlite database.
All tables are moved exactly as named using their existing column datatypes.
If a table alredy exists in the SQlite database an error will be returned.
This function is to be used to create new tables in a SQlite database rather
than append or overwrite them.
}
\details{
Microsoft Access 1997 or newer supported with Microsoft Access Driver (*.mdb, *.accdb) installed.
}
\keyword{access,}
\keyword{database}
\keyword{mdb,}
\keyword{sqlite,}
