% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owri_db.R
\name{owri_db}
\alias{owri_db}
\title{Current OWRI Database}
\usage{
owri_db(owri.db = "OwriDbExport_010826.db")
}
\arguments{
\item{owri.db}{Optional. owri SQLite database file name. Default is the
most recent version. 'OwriDbExport_010826.db'.}
}
\description{
Returns the path to the most recent SQLite OWRI database export included with this R package.
The original OWRI database published by the Oregon Watershed Enhancement Board can be found here:
https://hub.oregonexplorer.info/pages/water-planning-enhancing-watersheds-in-oregon#TheOWRIDatabaseandGISdata
}
\keyword{owri,}
\keyword{raw_data}
