% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owri_summary.R
\name{owri_summary}
\alias{owri_summary}
\title{Summarize OWRI treatments.}
\usage{
owri_summary(owri.db, complete.years = NA, huc8 = NA)
}
\arguments{
\item{owri.db}{The path and file name of the owri SQLite database.}

\item{complete.years}{Vector of numeric years used to fetch projects. Only projects completed in these years will be summarized. Default is NA resulting in summary of all years in the database.}

\item{huc8}{Vector of numeric HUC8 codes used to fetch projects. Only projects within the HUC8 will be summarized. Default is NA resulting in summary of all HUC8 codes in the database.}
}
\value{
Dataframe with the sum of treatments implemented grouped into five yearly periods by huc8 and activity type.
}
\description{
Summarize OWRI treatments.
}
\keyword{complete}
\keyword{huc8,}
\keyword{owri,}
\keyword{years,}
